requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose>=3.3.0
aiohttp>=3.9.0
cryptography>=42.0.0
python-multipart>=0.0.6
huggingface-hub>=0.21.4
selenium>=4.18.1
webdriver-manager>=4.0.1